+++ database/migrations/2023_02_23_095400_create_weight_genders_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->foreign('weight_id')->references('id')->on('weights')->onUpdate('cascade');
             $table->unsignedTinyInteger('gender_id')->unsigned();
             $table->foreign('gender_id')->references('id')->on('genders')->onUpdate('cascade');
-            $table->unsignedInteger('range_from');
-            $table->unsignedInteger('range_to');
+            $table->float('range_from');
+            $table->float('range_to');
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
         });
