+++ database/migrations/2023_02_23_105614_create_crcl_ranges_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->increments('id');
             $table->unsignedInteger('illness_category_id')->unsigned();
             $table->foreign('illness_category_id')->references('id')->on('illness_categories')->onUpdate('cascade');
-            $table->unsignedInteger('range_from');
-            $table->unsignedInteger('range_to');
+            $table->float('range_from');
+            $table->float('range_to');
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
         });
